+++ packages/s2-core/__tests__/unit/cell/col-cell-spec.ts
@@ -112,6 +112,8 @@ describe('Col Cell Tests', () => {
       id: 1,
       fieldValue: 'fieldValue',
       value: 'value',
+      width: 100,
+      height: 10,
     } as unknown as Node;
 
     const headerConfig: Partial<ColHeaderConfig> = {