+++ packages/s2-core/src/common/interface/basic.ts
@@ -414,3 +414,5 @@ export interface Point {
   x: number;
   y: number;
 }
+
+export type RowData = Data;