+++ packages/s2-core/src/common/interface/emitter.ts
@@ -1,4 +1,5 @@
 import type { FederatedPointerEvent as CanvasEvent } from '@antv/g';
+import type { SpreadSheet } from '../../sheet-type';
 import type { DataCell } from '../../cell/data-cell';
 import type { RowCell } from '../../cell/row-cell';
 import type { ColCell } from '../../cell/col-cell';
@@ -6,6 +7,7 @@ import type { S2Event } from '../../common/constant';
 import type {
   CellMeta,
   CellScrollPosition,
+  Data,
   HiddenColumnsInfo,
   LayoutResult,
   S2CellType,
@@ -71,8 +73,9 @@ export interface EmitterType {
 
   /** ================ Cell ================  */
   [S2Event.GLOBAL_LINK_FIELD_JUMP]: (data: {
+    cellData: Node;
     field: string;
-    record: RawData;
+    record: Data;
   }) => void;
 
   /** ================ Date Cell ================  */
@@ -139,6 +142,11 @@ export interface EmitterType {
     current: number;
   }) => void;
   [S2Event.LAYOUT_AFTER_HEADER_LAYOUT]: (data: LayoutResult) => void;
+  [S2Event.LAYOUT_AFTER_REAL_DATA_CELL_RENDER]: (options: {
+    add: [number, number][];
+    remove: [number, number][];
+    spreadsheet: SpreadSheet;
+  }) => void;
   /** @deprecated 请使用 S2Event.GLOBAL_SCROLL 代替 */
   [S2Event.LAYOUT_CELL_SCROLL]: (position: CellScrollPosition) => void;
   [S2Event.LAYOUT_CELL_MOUNTED]: (cell: S2CellType) => void;