+++ packages/s2-core/src/facet/base-facet.ts
@@ -72,9 +72,11 @@ import { CornerBBox } from './bbox/cornerBBox';
 import { PanelBBox } from './bbox/panelBBox';
 import {
   ColHeader,
+  type ColHeaderConfig,
   CornerHeader,
   Frame,
   RowHeader,
+  type RowHeaderConfig,
   SeriesNumberHeader,
 } from './header';
 import type { ViewCellHeights } from './layout/interface';
@@ -930,12 +932,12 @@ export abstract class BaseFacet {
   };
 
   /**
-    在当前表格滚动分两种情况:
-    1. 当前表格无滚动条: 无需阻止外部容器滚动
-    2. 当前表格有滚动条:
-      - 未滚动到顶部或底部: 当前表格滚动, 阻止外部容器滚动
-      - 滚动到顶部或底部: 恢复外部容器滚动
-  */
+   在当前表格滚动分两种情况:
+   1. 当前表格无滚动条: 无需阻止外部容器滚动
+   2. 当前表格有滚动条:
+   - 未滚动到顶部或底部: 当前表格滚动, 阻止外部容器滚动
+   - 滚动到顶部或底部: 恢复外部容器滚动
+   */
   isScrollOverTheViewport = (scrollOffset: CellScrollOffset) => {
     const { deltaY, deltaX, offsetY } = scrollOffset;
     const isScrollOverTheHeader = offsetY <= this.cornerBBox.maxY;
@@ -967,13 +969,13 @@ export abstract class BaseFacet {
   };
 
   /**
-    https://developer.mozilla.org/zh-CN/docs/Web/CSS/overscroll-behavior
-    阻止外部容器滚动: 表格是虚拟滚动, 这里按照标准模拟浏览器的 [overscroll-behavior] 实现
-    1. auto => 只有在滚动到表格顶部或底部时才触发外部容器滚动
-    1. contain => 默认的滚动边界行为不变（“触底”效果或者刷新），但是临近的滚动区域不会被滚动链影响到
-    2. none => 临近滚动区域不受到滚动链影响，而且默认的滚动到边界的表现也被阻止
-    所以只要不为 `auto`, 或者表格内, 都需要阻止外部容器滚动
-  */
+   https://developer.mozilla.org/zh-CN/docs/Web/CSS/overscroll-behavior
+   阻止外部容器滚动: 表格是虚拟滚动, 这里按照标准模拟浏览器的 [overscroll-behavior] 实现
+   1. auto => 只有在滚动到表格顶部或底部时才触发外部容器滚动
+   1. contain => 默认的滚动边界行为不变（“触底”效果或者刷新），但是临近的滚动区域不会被滚动链影响到
+   2. none => 临近滚动区域不受到滚动链影响，而且默认的滚动到边界的表现也被阻止
+   所以只要不为 `auto`, 或者表格内, 都需要阻止外部容器滚动
+   */
   private stopScrollChainingIfNeeded = (event: WheelEvent) => {
     const { interaction } = this.spreadsheet.options;
 
@@ -1141,6 +1143,11 @@ export abstract class BaseFacet {
       );
     });
     this.preCellIndexes = indexes;
+    this.spreadsheet.emit(S2Event.LAYOUT_AFTER_REAL_DATA_CELL_RENDER, {
+      add,
+      remove,
+      spreadsheet: this.spreadsheet,
+    });
   };
 
   protected init() {
@@ -1245,7 +1252,7 @@ export abstract class BaseFacet {
         position: { x: seriesNumberWidth, y },
         data: this.layoutResult.rowNodes,
         spreadsheet: this.spreadsheet,
-      });
+      } as unknown as RowHeaderConfig);
     }
     return this.rowHeader;
   }
@@ -1263,7 +1270,7 @@ export abstract class BaseFacet {
         data: this.layoutResult.colNodes,
         sortParam: this.spreadsheet.store.get('sortParam'),
         spreadsheet: this.spreadsheet,
-      });
+      } as unknown as ColHeaderConfig);
     }
     return this.columnHeader;
   }
@@ -1403,5 +1410,9 @@ export abstract class BaseFacet {
     return hiddenColumnsDetail.find((detail) =>
       detail.hideColumnNodes.some((node) => node.id === columnNode.id),
     );
+  }
+
+  public getCornerNodes(): Node[] {
+    return this.cornerHeader?.getNodes() || [];
   }
 }