+++ packages/s2-core/src/facet/header/corner.ts
@@ -133,6 +133,7 @@ export class CornerHeader extends BaseHeader<CornerHeaderConfig> {
     if (leafNode) {
       if (spreadsheet.isHierarchyTreeType()) {
         const cornerText = this.getTreeCornerText(options);
+
         const cornerNode: Node = new Node({
           id: '',
           field: '',
@@ -176,6 +177,7 @@ export class CornerHeader extends BaseHeader<CornerHeaderConfig> {
           cornerNode.spreadsheet = spreadsheet;
           cornerNodes.push(cornerNode);
         });
+        // spreadsheet type grid mode
       }
     }
 
@@ -279,5 +281,9 @@ export class CornerHeader extends BaseHeader<CornerHeaderConfig> {
         height,
       },
     });
+  }
+
+  public getNodes(): Node[] {
+    return this.headerConfig.data || [];
   }
 }