+++ packages/s2-core/src/facet/pivot-facet.ts
@@ -5,6 +5,7 @@ import {
   isArray,
   isEmpty,
   isNil,
+  isNumber,
   keys,
   last,
   map,
@@ -235,13 +236,13 @@ export class PivotFacet extends BaseFacet {
     const cellDraggedWidth = this.getColCellDraggedWidth(col);
 
     // 1. 拖拽后的宽度优先级最高
-    if (cellDraggedWidth) {
+    if (isNumber(cellDraggedWidth)) {
       return cellDraggedWidth;
     }
 
     // 2. 其次是自定义, 返回 null 则使用默认宽度
     const cellCustomWidth = this.getCellCustomSize(col, colCell?.width!);
-    if (!isNil(cellCustomWidth)) {
+    if (isNumber(cellCustomWidth)) {
       return cellCustomWidth;
     }
 
@@ -648,12 +649,12 @@ export class PivotFacet extends BaseFacet {
 
     const cellDraggedWidth = this.getRowCellDraggedWidth(node);
 
-    if (cellDraggedWidth) {
+    if (isNumber(cellDraggedWidth)) {
       return cellDraggedWidth;
     }
 
     const cellCustomWidth = this.getCellCustomSize(node, rowCell?.width!);
-    if (!isNil(cellCustomWidth)) {
+    if (isNumber(cellCustomWidth)) {
       return cellCustomWidth;
     }