+++ packages/s2-core/src/utils/interaction/select-event.ts
@@ -21,14 +21,14 @@ export const isMultiSelectionKey = (e: KeyboardEvent) => {
   );
 };
 
-export const getCellMeta = (cell: S2CellType) => {
+export const getCellMeta = (cell: S2CellType): CellMeta => {
   const meta = cell.getMeta();
-  const { id, colIndex, rowIndex } = meta;
-
+  const { id, colIndex, rowIndex, rowQuery } = meta;
   return {
     id,
     colIndex,
     rowIndex,
+    rowQuery,
     type: cell instanceof TableSeriesCell ? CellTypes.ROW_CELL : cell.cellType,
   };
 };
@@ -110,3 +110,70 @@ export function updateRowColCells(meta: ViewMeta) {
     });
   }
 }
+
+export const getRowHeaderByCellId = (
+  cellId: string,
+  s2: SpreadSheet,
+): Node[] => {
+  return s2.getRowNodes().filter((node: Node) => cellId.includes(node.id));
+};
+
+export const getColHeaderByCellId = (
+  cellId: string,
+  s2: SpreadSheet,
+): Node[] => {
+  return s2.getColumnNodes().filter((node: Node) => cellId.includes(node.id));
+};
+
+export const getInteractionCells = (
+  cell: CellMeta,
+  s2: SpreadSheet,
+): Array<CellMeta> => {
+  const { colHeader, rowHeader } = s2.interaction.getSelectedCellHighlight();
+
+  const headerGetters = [
+    {
+      shouldGet: rowHeader,
+      getter: getRowHeaderByCellId,
+    },
+    {
+      shouldGet: colHeader,
+      getter: getColHeaderByCellId,
+    },
+  ];
+
+  const selectedHeaderCells = headerGetters
+    .filter((item) => item.shouldGet)
+    .reduce((nodes, i) => [...nodes, ...i.getter(cell.id, s2)], [])
+    .filter((node) => !!node.belongsCell)
+    .map((node) => getCellMeta(node.belongsCell));
+
+  return [cell, ...selectedHeaderCells];
+};
+
+export const getInteractionCellsBySelectedCells = (
+  selectedCells: CellMeta[],
+  s2: SpreadSheet,
+): Array<CellMeta> => {
+  const headerSelectedCell: CellMeta[] = reduce(
+    selectedCells,
+    (_cells, selectedCell) => {
+      return [..._cells, ...getInteractionCells(selectedCell, s2)];
+    },
+    [],
+  );
+
+  // headerSelectedCell 会有重复的 cell，在这里统一去重
+  return uniqBy([...selectedCells, ...headerSelectedCell], 'id');
+};
+
+export const afterSelectDataCells = (root, updateDataCells) => {
+  const { colHeader, rowHeader } = root.getSelectedCellHighlight();
+  if (colHeader) {
+    root.updateCells(root.getAllColHeaderCells());
+  }
+  if (rowHeader) {
+    root.updateCells(root.getAllRowHeaderCells());
+  }
+  updateDataCells();
+};