+++ packages/s2-react/playground/config.ts
@@ -122,6 +122,7 @@ export const s2Options: SheetComponentOptions = {
       height: 50,
     },
     colCell: {
+      hideValue: true,
       width: 140,
       height: 30,
     },