+++ packages/s2-vue/src/utils/initPropAndEmits.ts
@@ -179,6 +179,7 @@ export const initBaseSheetEmits = () => {
     'scroll',
     'hover',
     // ============== Auto 自动生成的 ================
+    'layoutAfterRealDataCellRender',
     'rowCellBrushSelection',
     'colCellBrushSelection',
   ];